Create a file for each class
