+++ src/data_quality_utils/map_extractor/__init__.py
@@ -1,7 +1,5 @@
-from .map_extractor import (
-    GoogleMapTilesExtractor,
-    GoogleStaticMapsExtractor,
-    WMSExtractor,
-)
+from .google_static_maps_extractor import GoogleStaticMapsExtractor
+from .google_map_tiles_extractor import GoogleMapTilesExtractor
+from .wms_extractor import WMSExtractor
 
 __all__ = ["GoogleMapTilesExtractor", "GoogleStaticMapsExtractor", "WMSExtractor"]
